Refactor av kildekode, splittet opp til flere repo
