+++ src/main/java/servlets/ChangeKlubb.java
@@ -1,19 +1,14 @@
 package servlets;
 
 
-import models.UtoevereModel;
-import org.mariadb.jdbc.MariaDbDatabaseMetaData;
-import tools.repository.UserRepository;
+import tools.repository.KlubbRepo;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.xml.registry.infomodel.User;
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.util.List;
-import java.util.Set;
 
 
 @WebServlet(name = "ChangeKlubb", urlPatterns = {"/ChangeKlubb"})
@@ -46,7 +41,7 @@ public class ChangeKlubb extends AbstractAppServlet {
         String ki = req.getParameter("ki");
 
         if (action.contains("changeklubb")) {
-            UserRepository.changeKlubb(ck, ki, out);
+            KlubbRepo.changeKlubb(ck, ki, out);
             out.println("klubbens navn har blitt endret");
         }
