+++ src/main/java/servlets/RegistrerBruker.java
@@ -1,19 +1,14 @@
 package servlets;
 
 
-import models.UtoevereModel;
-import org.mariadb.jdbc.MariaDbDatabaseMetaData;
-import tools.repository.UserRepository;
+import tools.repository.LoginRepo;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.xml.registry.infomodel.User;
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.util.List;
-import java.util.Set;
 
 
 @WebServlet(name = "RegistrerBruker", urlPatterns = {"/RegistrerBruker"})
@@ -46,7 +41,7 @@ public class RegistrerBruker extends AbstractAppServlet {
         String pd = req.getParameter("pd");
 
         if (action.contains("Registrer")) {
-           UserRepository.Registrer(bn, pd, out);
+           LoginRepo.Registrer(bn, pd, out);
         }
 
     }
