+++ src/main/java/servlets/RemoveBruker.java
@@ -1,19 +1,14 @@
 package servlets;
 
 
-import models.UtoevereModel;
-import org.mariadb.jdbc.MariaDbDatabaseMetaData;
-import tools.repository.UserRepository;
+import tools.repository.UtoeverRepo;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.xml.registry.infomodel.User;
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.util.List;
-import java.util.Set;
 
 
 @WebServlet(name = "RemoveBruker", urlPatterns = {"/RemoveBruker"})
@@ -44,7 +39,7 @@ public class RemoveBruker extends AbstractAppServlet {
         String action = req.getParameter("action");
         String rm = req.getParameter("rm");
         if (action.contains("remove")) {
-            UserRepository.removeUtoever(rm, out);
+            UtoeverRepo.removeUtoever(rm, out);
         }
 
         }
