+++ src/main/java/servlets/RequestUserNameFromDb.java
@@ -1,14 +1,7 @@
 package servlets;
 
-import tools.DbTool;
-import tools.repository.UserRepository;
-
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
